+++ library/panic_unwind/src/lib.rs
@@ -92,8 +92,6 @@ extern "C" {
     fn __rust_foreign_exception() -> !;
 }
 
-mod dwarf;
-
 #[rustc_std_internal_symbol]
 #[allow(improper_ctypes_definitions)]
 pub unsafe extern "C" fn __rust_panic_cleanup(payload: *mut u8) -> *mut (dyn Any + Send + 'static) {
